backend/requirements.txt
fastapi
uvicorn
Pillow
tesserocr
